pyton/flask_backend/requirements.txt
flask
torch
transformers
pillow
